start migration to tsdx - rename files
