+++ .gitignore
@@ -1,6 +1,7 @@
-dist
-node_modules
-
-# logs & reports
-coverage
 *.log
+.DS_Store
+node_modules
+.rts2_cache_cjs
+.rts2_cache_esm
+.rts2_cache_umd
+dist
